requirements.txt
flask
flask_cors
flask_security_too
flask_sqlalchemy
flask_jwt_extended
Flask-Bcrypt
Flask-mail
python-dotenv
pillow
pymysql
pandas
datetime
openpyxl
ultralytics
tesseract
facenet_pytorch
ultralytics>=8.0.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
opencv-python>=4.5.0
flask>=2.0.0